requirements.txt
click==8.0.3
matplotlib==3.4.3
numpy==1.21.5
pandas==1.3.4
Pillow==9.0.1
Pympler==1.0.1
scikit_learn==1.0.2
scipy==1.7.1
tqdm==4.62.3
